Upated to export the MAX_MEM, USED_MEM and LAST_MON_TIME.

+++ AutoElastic/src/autoelastic/Graphic.java
@@ -105,7 +105,7 @@ public class Graphic {
         grafico.repaint();
     }
     
-    public void update(int contador, int cpu_utilizada, int cpu_disponivel, float threshold_max, float threshold_min){
+    public void update(int contador, float cpu_utilizada, float cpu_disponivel, float threshold_max, float threshold_min){
         dataset.addValue(cpu_utilizada, legendas[0], Integer.toString(contador));
         dataset.addValue(cpu_disponivel, legendas[1], Integer.toString(contador));
         dataset.addValue(threshold_max, legendas[2], Integer.toString(contador));
